Moving agents to a agents directory
