+++ .gitignore
@@ -1,5 +1,5 @@
 *.pth
-__pycache__/*
+__pycache__/
 .*.swp
 .*.swo
 .venv/
